Create README.md
Music reactive 12v RGB led strip is a fun project which changes its color based on the sound of the music. This project is built using an Arduino Uno board that controls multiple color effects of LED strips based on the frequency that hits the drum of the sound sensor. This RGB LED strip itself has multiple colors such as red, green, yellow, blue, orange, white, and, pink, which allows for a range of lighting effects. The sound sensor detects the beat of the music as input and changes the strip color and intensity based on the trigger created in the LED strip in response to the music or sound. The transistor TIP 122 is used to function the LED strip properly. With this project, one can create a unique dynamic light show and interactive visual display for events or parties.